[CHANGE] Repository manager

+++ DotBased.AspNet.Authority/Repositories/IAuthorityRepository.cs
@@ -2,6 +2,6 @@ namespace DotBased.AspNet.Authority.Repositories;
 
 public interface IAuthorityRepository
 {
-    public Task<int> GetVersion();
-    public Task SetVersion(int version);
+    public Task<long> GetVersion();
+    public Task SetVersion(long version);
 }
